server/requirements.txt
flask
selenium
youtube-transcript-api
undetected_chromedriver
langchain-text-splitters
langchain-google-genai
google-generativeai
python-dotenv
supabase
numpy

